Rename integration test files to match the method names being tested.
