+++ integration_tests/src/tests/prove_commit3_test.rs
